Class ClientTcpNMEA ready
